Bestätigungsdialog loeschen

+++ src/app/components/home/home.component.ts
@@ -4,6 +4,7 @@ import {EinkaufszettelActions} from "../../store/einkaufszettel/einkaufszettel.a
 import {selectAllEinkaufszettel} from "../../store/einkaufszettel/einkaufszettel.selectors";
 import {Artikel} from "../../entities/artikel";
 import {Einkaufszettel} from "../../entities/einkaufszettel";
+import {ConfirmationService, MessageService} from "primeng/api";
 
 @Component({
   selector: 'app-home',
@@ -13,7 +14,7 @@ import {Einkaufszettel} from "../../entities/einkaufszettel";
 export class HomeComponent implements OnInit {
   einkaufszettel!: Einkaufszettel[];
 
-  constructor(private store: Store) {
+  constructor(private store: Store, private confirmationService: ConfirmationService, private messageService: MessageService) {
   }
 
   ngOnInit(): void {
@@ -22,9 +23,20 @@ export class HomeComponent implements OnInit {
     this.store.select(selectAllEinkaufszettel).subscribe(einkaufszettel => this.einkaufszettel = JSON.parse(JSON.stringify(einkaufszettel))); // deep copy of store, so that changes are possible
   }
 
-  deleteArtikel(artikel: Artikel) {
-    this.store.dispatch(EinkaufszettelActions.deleteArtikel({data: artikel}));
-    this.store.dispatch(EinkaufszettelActions.loadEinkaufszettels());
+  deleteArtikel(event: Event, artikel: Artikel) {
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Sind Sie sich sicher, dass Sie den Artikel löschen möchten?',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      acceptIcon: "none",
+      rejectIcon: "none",
+      rejectButtonStyleClass: "p-button-text",
+      accept: () => {
+        this.store.dispatch(EinkaufszettelActions.deleteArtikel({data: artikel}));
+        this.store.dispatch(EinkaufszettelActions.loadEinkaufszettels());
+      }
+    });
   }
 
   changeArtikelGekauft(artikel: Artikel) {
